Change functions to return master keys
Also change receiver function to return a private key, which
can be used as master extended private key and derive addresses
to receive coins from a specific sender.

Change sender function to return a public key instead of
a serie of addresses. This public key is derived from the
receiver's private key mentioned before.

That way, the sender can create an xpub from that public key
and derive all addresses needed to send coins to the receiver.